Fix Navbar buttons + minor font issues (#82)
Ticket: NAUR-82
Navigation bar links only worked when directly clicking on the text. The
padding around the text does not, even when it should be.
MUI components which involve typography default to Arial, making any
text that still uses certain MUI components not use Roboto. This is
mainly an issue in the navigation bar and to a lesser extent, one part
of the footer. I've also added fallback fonts just in case.