Make markdown pages more flexible (#92)
Ticket: NAUR-94
This PR makes markdown pages more flexible to support custom slugs and
specify which files those custom slugs lead to.
`_meta.json` within the folders specified in `markdownFolders`
(`src/markdown/[folder]`) will be read to specify which pages need to be
generated by `npm run build`.
Example format:
```json
{
    "fru": {
        "index": "fru.mdx",
        "guide": {
            "index": "fru.mdx",
        }
    },
    "top": {
        "index": "top.mdx"
    },
    "dsr": {
        "index": "dsr.mdx"
    },
    "tea": {
        "index": "tea.mdx"
    },
    "uwu": {
        "index": "uwu.mdx"
    },
    "ucob": {
        "index": "ucob.mdx"
    }
}
```
Assuming this is within the folder `markdown/ultimates`, the slugs
generated will be:
```
/ultimates/fru/
/ultimates/fru/guide/
/ultimates/top/
/ultimates/dsr/
/ultimates/tea/
/ultimates/uwu/
/ultimates/ucob/
```
To specify which slug is a valid url, simply add a child with key
"index" and the file path as the value. File paths are relative to where
the `_meta.json` is located.
`_meta.json` which are located closer to the slug path has priority, so
for a slug like `/ultimates/dawntrail/dsr/guide/` with two `_meta.json`
in `markdown/ultimates` and another in `markdown/ultimates/dawntrail/`,
the latter will take priority if the specified slug exists in it.
Otherwise, it will fallback to the other one(s).

---------

Co-authored-by: Claire <nicholas.mellinger@outlook.com>