Add testing page (#135)
Ticket: NAUR-142
This adds a testing page that is only enabled when the `NEXT_PROD`
environment variable isn’t set to true. I have already configured the
environment variable on the main site, so the testing page should only
ever be available locally and on the staging domain.

Testing page is available at `/testing/page`, more can be added if
needed