Removed links to Findingway.io (#259)
This PR removes all links to Findingway.io. It was present in the
Desktop and Mobile menu, along with the footer 'Other Links'.
Findingway.io isn't functional at the moment and we are likely going to
fall back to using XIVPF for our data. Future development may want to
add the button to link to https://xivpf.com/listings.

Resolves #258

---------

Co-authored-by: Han <htra@wgu.edu>