[skip netlify] Add script to bulk compress images (#215)
It's pretty cumbersome to put images into Squoosh 1 by 1. Instead of
doing that all by hand, it would be faster to just use a script that
bulk compresses images that we want to compress instead.

Since we have a general idea of what settings we should aim for, using
the same settings for multiple images is fine. If one looks particularly
bad, we can always just change the settings for a specific image.

This script currently takes any `.png` file within the `uncompressed`
folder in the working directory when run (best to run this with a
working directory at the base of the project), compresses it, then
outputs it in the `compressed` folder with the .avif extension while
preserving folder structure.